[webgpu] Return kFatalFailure if ES 3.1 context creation fails
To gracefully disallow webpgu context, instead of crashing gpu process
later.

(cherry picked from commit eb16ea292048bdf9806766bf5f77b0576b6aacb7)

Bug: 483822628, 444049511
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7597668
Commit-Queue: Kramer Ge <fangzhoug@chromium.org>
Cr-Original-Commit-Position: refs/heads/main@{#1589436}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7624620
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Auto-Submit: Kramer Ge <fangzhoug@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/branch-heads/7680@{#1796}
Cr-Branched-From: 76b7d80e5cda23fe6537eed26d68c92e995c7f39-refs/heads/main@{#1582197}

+++ gpu/ipc/service/webgpu_command_buffer_stub.cc
@@ -55,7 +55,9 @@ WebGPUCommandBufferStub::WebGPUCommandBufferStub(
 
 WebGPUCommandBufferStub::~WebGPUCommandBufferStub() {
   // Must run before memory_tracker_ is destroyed.
-  decoder_context()->Destroy(false);
+  if (decoder_context()) {
+    decoder_context()->Destroy(false);
+  }
 
   memory_tracker_ = nullptr;
 }
